Make jest setup files share-able
