+++ jest/jest.fixture.setup.js
@@ -1,0 +1,42 @@
+/*
+Copyright 2019 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
+
+const eol = require('eol')
+const fs = require('fs')
+
+const fixturesFolder = './test/__fixtures'
+
+// helper for fixtures
+global.fixtureFile = (output) => {
+  return fs.readFileSync(`${fixturesFolder}/${output}`).toString()
+}
+
+// helper for fixtures
+global.fixtureJson = (output) => {
+  return JSON.parse(fs.readFileSync(`${fixturesFolder}/${output}`).toString())
+}
+
+// fixture matcher
+expect.extend({
+  toMatchFixture (received, argument) {
+    const val = fixtureFile(argument)
+    expect(eol.auto(received)).toEqual(eol.auto(val))
+    return { pass: true }
+  }
+})
+
+expect.extend({
+  toMatchFixtureJson (received, argument) {
+    const val = fixtureJson(argument)
+    expect(received).toEqual(val)
+    return { pass: true }
+  }
+})
