+++ jest/jest.setup.js
@@ -1,36 +1,20 @@
 /*
-Copyright 2019 Adobe. All rights reserved.
+Copyright 2019 Adobe Inc. All rights reserved.
 This file is licensed to you under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License. You may obtain a copy
 of the License at http://www.apache.org/licenses/LICENSE-2.0
+
 Unless required by applicable law or agreed to in writing, software distributed under
 the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
 
-module.exports = {
-  collectCoverage: true,
-  collectCoverageFrom: [
-    'src/**/*.js'
-  ],
-  coverageThreshold: {
-    global: {
-      branches: 100,
-      lines: 100,
-      statements: 100
-    }
-  },
-  reporters: [
-    'default',
-    'jest-junit'
-  ],
-  testEnvironment: 'node',
-  setupFilesAfterEnv: [
-    './jest/jest.setup.js',
-    './jest/jest.fetch.setup.js',
-    './jest/jest.fs.setup.js',
-    './jest/jest.swagger.setup.js',
-    './jest/jest.fixture.setup.js'
-  ]
-}
+const { stdout } = require('stdout-stderr')
+
+jest.setTimeout(30000)
+jest.useFakeTimers()
+
+// trap console log
+beforeEach(() => { stdout.start() })
+afterEach(() => { stdout.stop() })
