+++ jest/jest.swagger.setup.js
@@ -9,28 +9,5 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-module.exports = {
-  collectCoverage: true,
-  collectCoverageFrom: [
-    'src/**/*.js'
-  ],
-  coverageThreshold: {
-    global: {
-      branches: 100,
-      lines: 100,
-      statements: 100
-    }
-  },
-  reporters: [
-    'default',
-    'jest-junit'
-  ],
-  testEnvironment: 'node',
-  setupFilesAfterEnv: [
-    './jest/jest.setup.js',
-    './jest/jest.fetch.setup.js',
-    './jest/jest.fs.setup.js',
-    './jest/jest.swagger.setup.js',
-    './jest/jest.fixture.setup.js'
-  ]
-}
+// ensure a mocked openwhisk module for unit-tests
+jest.mock('swagger-client')
